Don't mix signed and unsigned.

+++ src/fullscr.c
@@ -1,7 +1,7 @@
 /* Copyright (C) 1997 DJ Delorie, see COPYING.DJ for details */
 /* Copyright (C) 1996 DJ Delorie, see COPYING.DJ for details */
 /* Copyright (C) 1995 DJ Delorie, see COPYING.DJ for details */
-/* $Id: fullscr.c,v 1.9 2001/12/12 07:32:07 pete Exp $ */
+/* $Id: fullscr.c,v 1.10 2002/03/30 06:31:12 pete Exp $ */
 /* ------------------------------------------------------------------------- */
 /*			    FULL SCREEN DEBUGGER			     */
 /*									     */
@@ -1501,7 +1501,7 @@ redraw (int first)
 	name = syms_val2name (breakpoint_table[b].addr, &delta);
 	if (name[0] != '0')
 	{
-	  if (delta && strlen (name) < width)
+	  if (delta && (int)strlen (name) < width)
 	    sprintf (buf, " %s+%#lx", name, delta);
 	  else
 	    sprintf (buf, " %-*s", width, name);
@@ -2069,7 +2069,7 @@ redraw (int first)
 	  if (evaluate (watch_pane_text[i], &val, &errtxt))
 	    {
 	      valtxt = errtxt;
-	      if (strlen (errtxt) > width - 5)
+	      if ((int)strlen (errtxt) > width - 5)
 		{
 		  valtxt = alloca (width - 4);
 		  strncpy (valtxt, errtxt, width - 4);
